core.common: move mime-related stuff to my.core.mime
no backward compat, unlikely it was used by anyone else
