+++ my/core/common.py
@@ -2,7 +2,6 @@ from glob import glob as do_glob
 from pathlib import Path
 from datetime import datetime
 from dataclasses import is_dataclass, asdict as dataclasses_asdict
-import functools
 import os
 from typing import (
     Any,
@@ -115,27 +114,6 @@ def get_files(
 
         paths = [wrap(p) for p in paths]
     return tuple(paths)
-
-
-@functools.lru_cache(1)
-def _magic():
-    import magic # type: ignore
-    return magic.Magic(mime=True)
-
-
-# TODO could reuse in pdf module?
-import mimetypes # todo do I need init()?
-# todo wtf? fastermime thinks it's mime is application/json even if the extension is xz??
-# whereas magic detects correctly: application/x-zstd and application/x-xz
-def fastermime(path: PathIsh) -> str:
-    paths = str(path)
-    # mimetypes is faster
-    (mime, _) = mimetypes.guess_type(paths)
-    if mime is not None:
-        return mime
-    # magic is slower but returns more stuff
-    # TODO Result type?; it's kinda racey, but perhaps better to let the caller decide?
-    return _magic().from_file(paths)
 
 
 Json = Dict[str, Any]
