+++ my/photos/main.py
@@ -15,9 +15,10 @@ from typing import Optional, NamedTuple, Iterator, Iterable, List
 
 from geopy.geocoders import Nominatim  # type: ignore
 
-from my.core.common import LazyLogger, fastermime
+from my.core import LazyLogger
 from my.core.error import Res, sort_res_by
 from my.core.cachew import cache_dir, mcachew
+from my.core.mime import fastermime
 
 from my.config import photos as config  # type: ignore[attr-defined]
